Fix: Critical - Relocate sensor async_setup_entry to sensor/__init__.py
This commit addresses the persistent failure of the sensor platform to load by moving the `async_setup_entry` function to the correct location and makes another attempt to fix the SSID client count SDK call.

- Sensor Platform Loading (`sensor/__init__.py`):
    - The `async_setup_entry` function, along with all necessary imports and logic for setting up sensor entities, has been moved from the (now deleted) `custom_components/meraki_ha/sensor.py` file to `custom_components/meraki_ha/sensor/__init__.py`.
    - Relative import paths within `sensor/__init__.py` have been adjusted to correctly load entity classes from the `sensor` directory (e.g., `from .device_status import ...`) and modules from the parent package (e.g., `from ..const import ...`).
    - This structural change is crucial and aligns with Home Assistant's platform loading mechanism when a directory named after the platform (e.g., "sensor") exists and is a package. It should resolve the `AttributeError: module 'custom_components.meraki_ha.sensor' has no attribute 'async_setup_entry'`.

- SSID Client Count SDK Call (`SSIDDeviceCoordinator`):
    - The SDK method call for fetching clients per network (to then filter by SSID) has been updated to `meraki_client.clients.get_network_clients(...)`. This is a further attempt to find the correct method on the `meraki.aio.AsyncDashboardAPI` instance after previous attempts (`.wireless.get_network_wireless_ssid_clients` and `.networks.get_network_clients`) failed with `AttributeError`.

These changes are vital for the integration's stability and core functionality, particularly for enabling all sensor entities, including the hardware device status sensors.
